+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/policies/application/rest/PolicyController.java
@@ -51,7 +51,7 @@ import java.util.Map;
 @Slf4j
 @RequiredArgsConstructor
 @RestController
-@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_SUPERVISOR')")
+@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
 @Tag(name = "Policies")
 @RequestMapping(path = "/policies")
 public class PolicyController {